% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{c2BroadSets}
\alias{c2BroadSets}
\title{This is data to be included in package}
\format{
GeneSetCollection
\describe{
\item{Genesetcollection}{GeneSetCollection from BroadCollection}
}
}
\usage{
c2BroadSets
}
\description{
This is data to be included in package
}
\keyword{datasets}
