% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{count_table}
\alias{count_table}
\title{This is data to be included in package}
\format{
An example count table where row names are gene ID, each column is a sample
\describe{
\item{counttable}{count table}
...
}
}
\usage{
count_table
}
\description{
This is data to be included in package
}
\keyword{datasets}
