% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.pathway.R
\name{dlPathwaysDB}
\alias{dlPathwaysDB}
\title{DL Pathways DB}
\usage{
dlPathwaysDB(pathway.db)
}
\arguments{
\item{pathway.db}{The databse to be used for encrichment analysis. Can be one of the following, "rWikiPathways", "KEGG", "REACTOME", "Hallmark","rWikiPathways_aug_2020"}
}
\value{
Returns a dataframe.
}
\description{
Download gene database for enrichment.
}
\references{
Xingpeng Li & Siddhartha Pachhai RVA - RNAseq Visualization Automation tool.
}
