% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_heatmap.R
\name{get.cpm.colors}
\alias{get.cpm.colors}
\title{Get CPM Colors}
\usage{
get.cpm.colors(data)
}
\arguments{
\item{data}{The CPM dataset.}
}
\description{
This function creates the color gradient for the cpm data.
}
