% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.cutoff.R
\name{ggsave.cutoff}
\alias{ggsave.cutoff}
\title{Create a message about fold change and pvalues used to produce the plot.}
\usage{
ggsave.cutoff(plot.save.to, gp)
}
\arguments{
\item{plot.save.to}{Path to where ggplot object from plot.cutoff will be saved.}

\item{gp}{This is the ggplot object from plot.cutoff}
}
\description{
This function processes summary statistics table generated by differential expression analysis
like \code{limma} or \code{DESeq2} and produces a message about pvalues and fold change used.
}
