% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_cutoff.R
\name{make.cutoff.plotly}
\alias{make.cutoff.plotly}
\title{Create plotly object for number of DE genes at different cutoff combinations}
\usage{
make.cutoff.plotly(df)
}
\arguments{
\item{df}{Summary statistics table from limma or DEseq2, where each row is a gene.}
}
\description{
This function processes summary statistics table generated by differential expression analysis
like \code{limma} or \code{DESeq2} to produce an interactibe visual object
which depicts the number of differntially expressed genes with different FDR and
fold change cutoff.
}
