% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_pathway.R
\name{multiPlot}
\alias{multiPlot}
\title{Multi Plot}
\usage{
multiPlot(allID, backup.d.sig, nd.res, ...)
}
\arguments{
\item{allID}{A vector of all pathway ID's from directional and non directional enriched datasets.}

\item{backup.d.sig}{A dataframe type of object with directional pathways data prior to any cutoff's being applied}

\item{nd.res}{A dataframe type of object with non directional pathways data prior to any cutoff's being applied}

\item{...}{pass on variables}
}
\value{
Returns ggplot.
}
\description{
Multi plot is for directional and non-directional plots
}
\details{
Multi plot is for directional and non-directional plots, when one of the plots doesn't contain observations.
}
\references{
Xingpeng Li & Siddhartha Pachhai RVA - RNAseq Visualization Automation tool.
}
