% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_pathway.R
\name{nullreturn}
\alias{nullreturn}
\title{Null Return}
\usage{
nullreturn(IS.list, type = 1)
}
\arguments{
\item{IS.list}{Indicator of whether the data frame being input is list or not.}

\item{type}{If type = 1(default) return directional null plot. If type = 2 return non directional null plot.}
}
\value{
The function returns either returns a data frame or the value NULL.
}
\description{
The function takes in a boolean value and a numeric value, which it uses to decide what to output.
}
\details{
nullreturn is a function that returns NULL for single df inputs that don't hold true for threshold values. It returns an empty dataframe for
list inputs which don't satisfy the cutoff's
}
\references{
Xingpeng Li & Siddhartha Pachhai RVA - RNAseq Visualization Automation tool.
}
