% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_pathway.R
\name{prettyGraphs}
\alias{prettyGraphs}
\title{Pretty Graphs}
\usage{
prettyGraphs(vizdf, ...)
}
\arguments{
\item{vizdf}{A dataframes of enriched pathways.}

\item{...}{pass on variables}
}
\value{
Returns a dataframe.
}
\description{
Special cases where list input and at least one treatment has signal but others don't.
}
\details{
Pretty Graphs is a function specifically meant to be in cases where one of the input treatments meet cutoff, but
one or more of the other treatments don't meet the cutoff values. This is important so that ggplot doesn't throw any errors.
}
\references{
Xingpeng Li & Siddhartha Pachhai RVA - RNAseq Visualization Automation tool.
}
