% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_cutoff.R
\name{produce.cutoff.warning}
\alias{produce.cutoff.warning}
\title{Create a warning about pvalue or FDR minimum value}
\usage{
produce.cutoff.warning(data, FDRflag)
}
\arguments{
\item{data}{Summary statistics table from limma or DEseq2, where each row is a gene.}

\item{FDRflag}{The column name of the False Discovery Rate (FDR) in the summary statistics table.}
}
\description{
This function processes summary statistics table generated by differential expression analysis
like \code{limma} or \code{DESeq2} and produces a warning about pvalue or FDR minimum value
}
