% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_gene.R, R/plot_heatmap.R
\name{reformat.ensembl}
\alias{reformat.ensembl}
\title{Reformat Ensembl GeneIDs}
\usage{
reformat.ensembl(logcpm, ct.table.id.type)

reformat.ensembl(logcpm, ct.table.id.type)
}
\arguments{
\item{logcpm}{The input count table transformed into log counts per million.}

\item{ct.table.id.type}{The gene id format in \code{logcpm} should be one of: ACCNUM, ALIAS, ENSEMBL, ENSEMBLPROT,
ENSEMBLTRANS, ENTREZID, ENZYME, EVIDENCE, EVIDENCEALL, GENENAME, GO, GOALL, IPI, MAP, OMIM,
ONTOLOGY, ONTOLOGYALL, PATH, PFAM, PMID, PROSITE, REFSEQ, SYMBOL, UCSCKG, UNIGENE, UNIPROT.}
}
\description{
This is the function to exclude the version number from the input ensembl type gene ids.

This is the function to exclude the version number from the input ensembl type gene ids.
}
