% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sample_annotation}
\alias{sample_annotation}
\title{This is data to be included in package}
\format{
Sample annotation document
\describe{
\item{sample_id}{sample name}
\item{tissue}{tissue for comparison}
\item{subject_id}{subject id}
\item{day}{time points}
...
}
}
\usage{
sample_annotation
}
\description{
This is data to be included in package
}
\keyword{datasets}
