% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sample_count_cpm}
\alias{sample_count_cpm}
\title{This is data to be included in package}
\format{
An example cpm table where row names are gene ID, each column is a sample
\describe{
\item{counttable}{count cpm table}
...
}
}
\usage{
sample_count_cpm
}
\description{
This is data to be included in package
}
\keyword{datasets}
