% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Sample_disease_gene_set}
\alias{Sample_disease_gene_set}
\title{This is data to be included in package}
\format{
An example disease gene set from summary statistics table as dataframe, row names are gene ID
the summary statistics can be calculated from disease vs healthy, which is this example.
\describe{
\item{logFC}{log2 fold change from comparison}
\item{AveExpr}{Average expression for this gene}
\item{P.Value}{p value}
\item{adj.P.Val}{adjusted p value or FDR}
...
}
}
\usage{
Sample_disease_gene_set
}
\description{
This is data to be included in package
}
\keyword{datasets}
