% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Sample_summary_statistics_table1}
\alias{Sample_summary_statistics_table1}
\title{This is data to be included in package}
\format{
Second example summary statistics table as dataframe, row names are gene ID
\describe{
\item{logFC}{log2 fold change from comparison}
\item{AveExpr}{Average expression for this gene}
\item{P.Value}{p value}
\item{adj.P.Val}{adjusted p value or FDR}
...
}
}
\usage{
Sample_summary_statistics_table1
}
\description{
This is data to be included in package
}
\keyword{datasets}
