% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_pathway.R
\name{secondCutoffErr}
\alias{secondCutoffErr}
\title{Second Cutoff Error}
\usage{
secondCutoffErr(df, comp.names, TypeQ = 1)
}
\arguments{
\item{df}{A list of dataframes.}

\item{comp.names}{a character vector contain the comparison names corresponding to the same order to the \code{dat.list}. default = NULL.}

\item{TypeQ}{If type = 1(default) return directional null plot. If type = 2 return non directional null plot.}
}
\value{
Returns a dataframe.
}
\description{
The function takes in a list of dataframe, comp names and a specified type, to output a dataframe styled for ggplot.
}
\details{
secondCutoffErr is a function specifically meant to be used for list inputs. It is used for cases where after
applying filter to the data, one of the comparison ID gets left out, this adversely effects the ggplot
}
\references{
Xingpeng Li & Siddhartha Pachhai RVA - RNAseq Visualization Automation tool.
}
