% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_gene.R, R/plot_heatmap.R
\name{transform.geneid}
\alias{transform.geneid}
\title{Transform GeneIDs}
\usage{
\method{transform}{geneid}(gene.names, from = ~gene.id.type, to = ~ct.table.id.type)

\method{transform}{geneid}(gene.names, from = ~gene.id.type, to = ~ct.table.id.type)
}
\arguments{
\item{gene.names}{Genes,in the format of character vector, to be transformed.}

\item{from}{The gene id format of \code{gene.names}, should be one of: ACCNUM, ALIAS, ENSEMBL, ENSEMBLPROT,
ENSEMBLTRANS, ENTREZID, ENZYME, EVIDENCE, EVIDENCEALL, GENENAME, GO, GOALL, IPI, MAP, OMIM,
ONTOLOGY, ONTOLOGYALL, PATH, PFAM, PMID, PROSITE, REFSEQ, SYMBOL, UCSCKG, UNIGENE, UNIPROT.}

\item{to}{The new gene id format should be one of: ACCNUM, ALIAS, ENSEMBL, ENSEMBLPROT,
ENSEMBLTRANS, ENTREZID, ENZYME, EVIDENCE, EVIDENCEALL, GENENAME, GO, GOALL, IPI, MAP, OMIM,
ONTOLOGY, ONTOLOGYALL, PATH, PFAM, PMID, PROSITE, REFSEQ, SYMBOL, UCSCKG, UNIGENE, UNIPROT.}
}
\description{
This is the function to transform the input gene id type to another gene id type.

This is the function to transform the input gene id type to another gene id type.
}
