% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.input.R
\name{validate.baseline}
\alias{validate.baseline}
\title{Validate Baseline Values}
\usage{
validate.baseline(annot, baseline.val, baseline.flag)
}
\arguments{
\item{annot}{The annotation dataframe.}

\item{baseline.val}{The baseline value passed by the user.}

\item{baseline.flag}{The baseline.flag passed by the user.}
}
\description{
Ensures that user-input \code{baseline.val} and \code{baseline.flag}
parameters are valid with respect to the \code{annot} dataframe.
}
\details{
Specifically, validates that \code{baseline.flag} value(s) are columns
in \code{annot}, and that \code{baseline.val} value(s) occur at least once in
their respective \code{baseline.flag} columns.
}
\references{
Xingpeng Li, Tatiana Gelaf Romer & Siddhartha Pachhai RVA - RNAseq Visualization Automation tool.
}
