% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.input.R
\name{validate.col.types}
\alias{validate.col.types}
\title{Check Summary Statistics Required Column Types}
\usage{
validate.col.types(datin, name = 1, flags)
}
\arguments{
\item{datin}{the summary statistics file.}

\item{name}{summary statistics file position indicator}

\item{flags}{\code{FCflag} or \code{FDRflag} to be checked}
}
\description{
\code{FCflag} and \code{FDRflag} must be numeric.
}
\references{
Xingpeng Li, Tatiana Gelaf Romer & Siddhartha Pachhai RVA - RNAseq Visualization Automation tool.
}
