% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.input.R
\name{validate.comp.names}
\alias{validate.comp.names}
\title{Validate Comp Names}
\usage{
validate.comp.names(comp.names, data)
}
\arguments{
\item{comp.names}{a character vector contain the comparison names corresponding to the same order to the \code{dat.list}. default = NULL.}

\item{data}{summary statistics table (data.frame) from limma or DEseq2, where rownames are gene id.}
}
\description{
This function ensures that when a list of data frames are used as input the
the number of comp names are the same as the number of data frames.
}
\references{
Xingpeng Li, Tatiana Gelaf Romer & Siddhartha Pachhai RVA - RNAseq Visualization Automation tool.
}
