% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.input.R
\name{validate.data.annot}
\alias{validate.data.annot}
\title{Validate Data in the Context of Annotation}
\usage{
validate.data.annot(data, annot, sample.id)
}
\arguments{
\item{data}{input data}

\item{annot}{annotation file}

\item{sample.id}{sample id in the input}
}
\description{
Ensures that the annotation file matches the data file with
respect to sample IDs. Throws warnings if there are discrepencies.
}
\references{
Xingpeng Li, Tatiana Gelaf Romer & Siddhartha Pachhai RVA - RNAseq Visualization Automation tool.
}
