% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.input.R
\name{validate.flag}
\alias{validate.flag}
\title{Validate Flag Value Is Valid}
\usage{
validate.flag(value, name, Options)
}
\arguments{
\item{value}{The user-input value for the parameter}

\item{name}{The name of the parameter to be displayed in the error}

\item{Options}{A vector of valid values for \code{value}}
}
\description{
Enures that the \code{value} is one of \code{Options} and throws an error
otherwise.
}
\references{
Xingpeng Li, Tatiana Gelaf Romer & Siddhartha Pachhai RVA - RNAseq Visualization Automation tool.
}
