% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.input.R
\name{validate.genes.present}
\alias{validate.genes.present}
\title{Validate genes present}
\usage{
validate.genes.present(data.genes, geneset)
}
\arguments{
\item{data.genes}{The gene id's.}

\item{geneset}{a summary statistic table contain the genes want to be highlighted, the gene name format (in row names)
needs to be consistent to the main summary statistics table). For example, this summary statistics
table coulb be the output summary statistics table from Disease vs Healthy comparison (Only contain
the subsetted significant genes want to be highlighted).}
}
\description{
Checks how many of the gene id's in the dataset are there in the geneset.
}
\references{
Xingpeng Li, Tatiana Gelaf Romer & Siddhartha Pachhai RVA - RNAseq Visualization Automation tool.
}
