% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.input.R
\name{validate.numeric}
\alias{validate.numeric}
\title{Validate Numeric Column}
\usage{
validate.numeric(datin, col, name = 1)
}
\arguments{
\item{datin}{The data in question.}

\item{col}{The column to validate as numeric.}

\item{name}{the position of dataset}
}
\description{
Ensures that a column in a dataframe which must be numeric is
numeric and throws an error otherwise.
}
\details{
This specifically checks if any of the values in the column can
be coerced as numeric.
}
\references{
Xingpeng Li, Tatiana Gelaf Romer & Siddhartha Pachhai RVA - RNAseq Visualization Automation tool.
}
