% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.input.R
\name{validate.pval.range}
\alias{validate.pval.range}
\title{Validate P-value Range}
\usage{
validate.pval.range(pval, name)
}
\arguments{
\item{pval}{The pvalue}

\item{name}{The name of the value to include in the error.}
}
\description{
Error-handling for invalid p-value.
}
\references{
Xingpeng Li, Tatiana Gelaf Romer & Siddhartha Pachhai RVA - RNAseq Visualization Automation tool.
}
