% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.input.R
\name{validate.pvals}
\alias{validate.pvals}
\title{Validate Pvalues}
\usage{
validate.pvals(p.min, p.max, p.step)
}
\description{
This function ensures the fold change minimum, maximum, and step
are valid.
}
\details{
Specifically it checks that the pvalues are between 0-1, and that
at least 1 \code{p.step} fits within the \code{p.min} and \code{p.max} bounds and
is positive.
}
\references{
Xingpeng Li, Tatiana Gelaf Romer & Siddhartha Pachhai RVA - RNAseq Visualization Automation tool.
}
