% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.input.R
\name{validate.single.table.isnotlist}
\alias{validate.single.table.isnotlist}
\title{Validate Single Table is not list}
\usage{
validate.single.table.isnotlist(data)
}
\arguments{
\item{data}{summary statistics table (data.frame) from limma or DEseq2, where rownames are gene id.}
}
\description{
Makes sure the summary table being input is of the right class and format.
}
\references{
Xingpeng Li, Tatiana Gelaf Romer & Siddhartha Pachhai RVA - RNAseq Visualization Automation tool.
}
