% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.input.R
\name{validate.stats.cols}
\alias{validate.stats.cols}
\title{Check Summary Statistics Required Columns}
\usage{
validate.stats.cols(datin, name = 1, req.cols)
}
\arguments{
\item{datin}{the summary statistics file.}

\item{name}{summary statistics file position indicator}

\item{req.cols}{required column names of \code{FCflag} and \code{FDRflag} pass on from validate.stats}
}
\description{
Required columns are \code{FCflag} and \code{FDRflag}
}
\references{
Xingpeng Li, Tatiana Gelaf Romer & Siddhartha Pachhai RVA - RNAseq Visualization Automation tool.
}
