% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.input.R
\name{validate.stats}
\alias{validate.stats}
\title{Validate Summary Statistics File}
\usage{
validate.stats(datin, name = 1, ...)
}
\arguments{
\item{datin}{the summary statistics file.}

\item{name}{summary statistics file position indicator}

\item{...}{pass on variables}
}
\description{
Check for required column names and types.
}
\references{
Xingpeng Li, Tatiana Gelaf Romer & Siddhartha Pachhai RVA - RNAseq Visualization Automation tool.
}
