% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{wpA2020}
\alias{wpA2020}
\title{This is data to be included in package}
\format{
Rwikipathway data downloaded version 2020
\describe{
\item{name}{pathway name}
\item{version}{version}
\item{wpid}{pathway id}
\item{org}{host name}
...
}
}
\usage{
wpA2020
}
\description{
This is data to be included in package
}
\keyword{datasets}
