% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_heatmap.R
\name{plot_heatmap.cpm}
\alias{plot_heatmap.cpm}
\title{Plot a CPM Heatmap}
\usage{
plot_heatmap.cpm(cpm, annot, title = "RVA CPM Heatmap", ...)
}
\arguments{
\item{cpm}{cpm data}

\item{annot}{A long-format dataframe with any pertinent treatment data about
the samples. The only required column is one titled the \code{sample.id}
value with values matching the column names of sample IDs in \code{data}.
Additional columns can contain information such as treament
compounds, dates of sample collection, or dosage quantities.}

\item{title}{A title for the heatmap. Default = "RVA Heatmap".}

\item{...}{pass over parameters}
}
\description{
An alias for \code{plot_heatmap.expr(annot, cpm, fill = "CPM", ...)}.
}
